initialized game state
Had to remove references from that function so it would work within the eGUI method.

Now that the game state is initialized we can update the GUI elements like the field player labels and stuff.

First need to create a function to find a player by position within a given roster.